+++ LookupAnything/Framework/Fields/ItemIconListField.cs
@@ -14,6 +14,9 @@ internal class ItemIconListField : GenericField
     /*********
     ** Fields
     *********/
+    /// <summary>The text to show before the item list, if any.</summary>
+    private readonly string? IntroText;
+
     /// <summary>The items to draw.</summary>
     private readonly Tuple<Item, SpriteInfo?>[] Items;
 
@@ -22,6 +25,9 @@ internal class ItemIconListField : GenericField
 
     /// <summary>Whether to draw the stack size on the item icon.</summary>
     private readonly bool ShowStackSize;
+
+    /// <summary>The pixel indent to apply before each entry in the list.</summary>
+    private readonly int IconIndent = 0;
 
 
     /*********
@@ -32,14 +38,17 @@ internal class ItemIconListField : GenericField
     /// <param name="label">A short field label.</param>
     /// <param name="items">The items to display.</param>
     /// <param name="showStackSize">Whether to draw the stack size on the item icon.</param>
+    /// <param name="introText">The text to show before the item list, if any.</param>
     /// <param name="formatItemName">Get the name to show for an item, or <c>null</c> to use the item's display name.</param>
-    public ItemIconListField(GameHelper gameHelper, string label, IEnumerable<Item?>? items, bool showStackSize, Func<Item, string?>? formatItemName = null)
+    public ItemIconListField(GameHelper gameHelper, string label, IEnumerable<Item?>? items, bool showStackSize, string? introText = null, Func<Item, string?>? formatItemName = null, int iconIndent = 0)
         : base(label, hasValue: items != null)
     {
         this.Items = items?.WhereNotNull().Select(item => Tuple.Create(item, gameHelper.GetSprite(item))).ToArray() ?? [];
         this.HasValue = this.Items.Any();
         this.ShowStackSize = showStackSize;
+        this.IntroText = introText;
         this.FormatItemName = formatItemName;
+        this.IconIndent = iconIndent;
     }
 
     /// <inheritdoc />
@@ -49,23 +58,31 @@ internal class ItemIconListField : GenericField
         float textHeight = font.MeasureString("ABC").Y;
         Vector2 iconSize = new Vector2(textHeight);
 
+        // draw intro
+        int topOffset = 0;
+        if (this.IntroText != null)
+        {
+            Vector2 textSize = spriteBatch.DrawTextBlock(font, this.IntroText, position, wrapWidth);
+            topOffset += (int)Math.Max(iconSize.Y, textSize.Y) + 10;
+        }
+
         // draw list
         const int padding = 5;
-        int topOffset = 0;
+        int leftOffset = this.IconIndent;
         foreach ((Item item, SpriteInfo? sprite) in this.Items)
         {
             // draw icon
-            spriteBatch.DrawSpriteWithin(sprite, position.X, position.Y + topOffset, iconSize);
+            spriteBatch.DrawSpriteWithin(sprite, position.X + leftOffset, position.Y + topOffset, iconSize);
             if (this.ShowStackSize && item.Stack > 1)
             {
                 float scale = 2f; //sprite.SourceRectangle.Width / iconSize.X;
-                Vector2 sizePos = position + new Vector2(iconSize.X - Utility.getWidthOfTinyDigitString(item.Stack, scale), iconSize.Y + topOffset - 6f * scale);
+                Vector2 sizePos = position + new Vector2(leftOffset + iconSize.X - Utility.getWidthOfTinyDigitString(item.Stack, scale), iconSize.Y + topOffset - 6f * scale);
                 Utility.drawTinyDigits(item.Stack, spriteBatch, sizePos, scale: scale, layerDepth: 1f, Color.White);
             }
 
             // draw text
             string displayText = this.FormatItemName?.Invoke(item) ?? item.DisplayName;
-            Vector2 textSize = spriteBatch.DrawTextBlock(font, displayText, position + new Vector2(iconSize.X + padding, topOffset), wrapWidth);
+            Vector2 textSize = spriteBatch.DrawTextBlock(font, displayText, position + new Vector2(leftOffset + iconSize.X + padding, topOffset), wrapWidth);
 
             topOffset += (int)Math.Max(iconSize.Y, textSize.Y) + padding;
         }