+++ LookupAnything/docs/release-notes.md
@@ -5,6 +5,7 @@
 * Added controller snap navigation for the search UI.
 * Added experimental screen reader support for the search UI.
 * Added RGB format to color data mining fields.
+* Added achievement lookups to see basic achievement info.
 * Fixed search UI in 1.53.0 deselecting the search box when the cursor isn't over it.
 * Fixed support for crops with multiple fertilizers applied via Ultimate Fertilizer.
 * Improved translations. Thanks to JanUwU42 (updated German), MakinDay (updated Italian), mehmetgorkemarslan (updated Turkish), and SummerFleur2997 (updated Chinese)!